+++ ogc-usb-hid/client.h
@@ -1,0 +1,10 @@
+#ifndef OGC_USB_HID_CLIENT_H
+#define OGC_USB_HID_CLIENT_H
+
+#include <stdbool.h>
+
+#define MAX_FAKE_WIIMOTES 2
+
+void ogc_usb_hid_set_sensor_bar_position_top(bool on_top);
+
+#endif
