+++ ogc-usb-hid/input_device.h
