+++ ogc-usb-hid/internals.h
@@ -1,0 +1,8 @@
+#ifndef OGC_USB_HID_INTERNALS_H
+#define OGC_USB_HID_INTERNALS_H
+
+#include <stdbool.h>
+
+extern bool _usb_hid_sensor_bar_position_top;
+
+#endif
