+++ ogc-usb-hid/usb_device_drivers.h
@@ -22,6 +22,15 @@ static inline bool usb_driver_is_compatible(u16 vid, u16 pid, const struct devic
     return false;
 }
 
+/* API for device drivers to receive and report controller events */
+void fake_wiimote_set_extension(fake_wiimote_t *wiimote, enum wiimote_ext_e ext);
+void fake_wiimote_report_input(fake_wiimote_t *wiimote, u16 buttons);
+void fake_wiimote_report_accelerometer(fake_wiimote_t *wiimote, u16 acc_x, u16 acc_y, u16 acc_z);
+void fake_wiimote_report_ir_dots(fake_wiimote_t *wiimote,
+                                 struct ir_dot_t ir_dots[static IR_MAX_DOTS]);
+void fake_wiimote_report_input_ext(fake_wiimote_t *wiimote, u16 buttons, const void *ext_data,
+                                   u8 ext_size);
+
 extern const usb_device_driver_t ds3_usb_device_driver;
 extern const usb_device_driver_t ds4_usb_device_driver;
 extern const usb_device_driver_t dr_usb_device_driver;
