+++ ogc-usb-hid/usb_drivers/sony_ds3.c
